Re-add vehicle prefab spawning
