+++ src/testapp/sessions/physics.h
@@ -37,23 +37,6 @@
 namespace testapp::scenes
 {
 
-struct SpawnShape
-{
-    osp::Vector3    m_position;
-    osp::Vector3    m_velocity;
-    osp::Vector3    m_size;
-    float           m_mass;
-    osp::EShape     m_shape;
-};
-
-struct ACtxPhysShapes
-{
-    osp::active::ActiveEntSet_t     ownedEnts;
-
-    std::vector<SpawnShape>         m_spawnRequest;
-    osp::active::ActiveEntVec_t     m_ents;
-    osp::draw::MaterialId           m_materialId;
-};
 
 /**
  * @brief Physical properties for entities and generic Physics interface
@@ -65,26 +48,6 @@ osp::Session setup_physics(
         osp::ArrayView<entt::any>   topData,
         osp::Session const&         scene,
         osp::Session const&         commonScene);
-
-/**
- * @brief Queues and logic for spawning physics shapes
- */
-osp::Session setup_phys_shapes(
-        osp::TopTaskBuilder&        rBuilder,
-        osp::ArrayView<entt::any>   topData,
-        osp::Session const&         scene,
-        osp::Session const&         commonScene,
-        osp::Session const&         physics,
-        osp::draw::MaterialId       materialId);
-
-osp::Session setup_phys_shapes_draw(
-        osp::TopTaskBuilder&        rBuilder,
-        osp::ArrayView<entt::any>   topData,
-        osp::Session const&         windowApp,
-        osp::Session const&         sceneRenderer,
-        osp::Session const&         commonScene,
-        osp::Session const&         physics,
-        osp::Session const&         physShapes);
 
 /**
  * @brief Queues and logic for spawning Prefab resources
